Set service's self url based on settings for Horizon
The intention of the original patch is to remove Horizon's reference to
internal endpoint so that the default endpoint type specified in settings
(OPENSTACK_ENDPOINT_TYPE) can take effect uniformly across all services.
But the patched function base.get_url_for_service() take 3 mandatory
parameters. So removing the third param causes Horizon traceback when
accessing API Access page.

This commit updated the patch so that services's self url is set to
the value of OPENSTACK_ENDPOINT_TYPE in settings (or default to
"internalURL" if OPENSTACK_ENDPOINT_TYPE is not set).

Change-Id: I19604d702f5f2de80dc4d0f426154fb5a0b9ad78
Closes-Bug: 1874902
Signed-off-by: Andy Ning <andy.ning@windriver.com>