Update cinder directives build file
Currently, cinder docker image doesn't contain nfs mount helpers
installed, so trying to mount a NFS volume on a cinder-backup
container fails.
In order to enable support for NFS backend on cinder-backup, we
need to install the nfs-utils package on cinder image, so that
cinder-backup is able to mount NFS volumes.

Task: 41796
Story: 2008613
Change-Id: Ib8e4675069292dc43f98ff55c25626a19ed37b12
Signed-off-by: Daniel Pereira <Daniel.Pereira@windriver.com>