Remove dbconfig and openstack-pkg-tools config
Remove the dbconfig and openstack-pkg-tools post configuration
since we use puppet to configure the services and doing
both will lead the problems with integration.

The patch is formatted like this since the debian/rules
found in the git repository doesnt exactly match up
with what is in the Debian archive.

Story: 2009101
Task: 44026

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I7fc28ba871aee694eed94ddac6729c4c14d31aad