Python various packages: change dl_name to dl_path
Fix a typo in meta_data.yaml. The field name should
be "dl_path", not "dl_name".

If debrepack detects none of the "src_path", "dl_path"
and "dl_hook" in meta_data.yaml, it considers the package
is a Debian native package, and uses "apt source" to
download it. The 3 packages are just the Debian native
packages, so no error before, but "apt source" is not
expected.

Test Plan:
Pass: successfully build test

Story: 2009101
Task: 43518

Signed-off-by: Yue Tao <yue.tao@windriver.com>
Change-Id: I066a7fee8f6523ccd6ba945925c5c7089a6bfeb2