Fix old kubectl client in stx-platformclients (from 1.5.2 to 1.21.3)
The actual kubernetes-client is in a much older version due to a
diferent version that is installed in the docker image.

As the new client change its name for kubernetes-1.21.3-client, the
version we are downloading is a older release that is downloaded
from the "extras" repository in the CentOS 7.

This commit fix the package name and version. The problem is that the
rpm spec change the default binary location to:
/usr/local/kubernetes/1.21.3/stage2/usr/bin/kubectl.

So to fix this we still have to link kubectl to /usr/bin using the
CUSTOMIZATION variable.

Partial-Bug: #1955427

Signed-off-by: Thiago Miranda <ThiagoOliveira.Miranda@windriver.com>
Change-Id: If68f397b9ced685b9a2631057f2dd0aca3667709