Setting PYTHON3=no for LOCI images
Recently, LOCI changed its default so images are python3 if nothing is
specified at the *docker_image files. With the recent upversion of LOCI
in the build tools [1], we need to explictly set PYTHON3=no to images
that are still using python2. This commit fixes the StarlingX images to
use this parameter.

This will also remove all PYTHON3=yes since it's the detault now to
reduce uneeded LOCs and ease maintenance.

TEST PLAN
PASS build the stx-nova on the stable branch
PASS build the stx-openstackclients on the stable branch
PASS build the stx-platformclients on the stable branch

[1] https://opendev.org/starlingx/root/commit/00bd632a68b8172f5d4baec2b1e4f8a08b9d5be5

Closes-Bug: #1958696
Depends-On: https://review.opendev.org/c/starlingx/root/+/826155
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: Idffec6b57371326e8667ac4b16ac9b11f22afe1a