First update on image customization
Following the openstack transition to Debian, this aims to start the
creation of the debian docker image files, adapting the customization
to proper enable WSGI on Debian.

Since we are not currently building Openstack on debian yet, this
won't have any effect on our current build/testing.

Also, since these images use apache, there's a change which is yet to
be done on the stx-openstack-helm manifest to user the "www-data"
socket-user insetad of "apache", since the latter one is for centos
only.

Test Plan:

PASS: Build horizon and placement images using a debian base image
PASS: Override both images on a working Openstack application
PASS: Access the horizon interface
PASS: Remove and apply Openstack

Story: 2010072
Task: 45558

Signed-off-by: Pedro Almeida <pedro.monteiroazevedodemouraalmeida@windriver.com>
Change-Id: I776e03b863056fbb068e2eca0637e9c8b64b4b0c