Change version requirement of netaddr
Netaddr is using version 0.7.19 on remote cli. Because of this
version, commands being executed towards platform apis are causing a
syntax warning.
This fix changes the minimum requirement of netaddr to 0.7.20 to
address a fix for this warning.

Test Plan:
PASS: Build debian iso and perform fresh install.
PASS Build python3 wheels tarball on Debian and build
stx-platformclients image on Debian.


Closes-Bug: 1999563

Signed-off-by: Luiz Felipe Kina <LuizFelipe.EiskeKina@windriver.com>
Change-Id: I0c111875f622be9696bf143b321f1a8dfd594c7f