Add PROJECT_UID and PROJECT_GID specification to stx-cinder
This review intends to solve a id conflict between the
cinder user and the keystone user (added on [1]). The
keystone user is attached with the id 42424, same as
the cinder user. This conflict is making volume related
commands to fail due permissions of the user (because it
is trying to execute the commands as keystone user, not
cinder).

[1] https://review.opendev.org/c/starlingx/integ/+/854246

Test plan:
PASS - Check the /etc/passwd file to see that the user
       'cinder' in the cinder container changed its id
       from 42424 to 42425.

Partial-Bug: 2012392

Signed-off-by: Rafael Falcao <rafael.vieirafalcao@windriver.com>
Change-Id: I29bba77beb0e63dfd03fcc681aba8a13b4c3445f