Add support for connect and read timeouts
The current keystone auth library only allows the configuration of 1
timeout value, that gets used for both connection_timeout and
read_timeout in the lower layers of the python requests library.

This change enables the usage of both connection and read timeout
when creating keystone session. This change is localized to session.py
because it's already supported by the python requests library.

It will enable for fail fast when the endpoint is offline when used
with shorter values for connect timeout while keeping a longer
read_timeout.

For instance: (5, 30), where (connection_timeout, read_timeout)

This change is to support task 50365 in [1].
[1] https://storyboard.openstack.org/#!/story/2011106

Test plan:

PASS: Run a full build, install, bootstrap and unlock.
PASS: Check alarms with 'fm alarm-list'

Story: 2011106
Task: 50363

Change-Id: Iab5c784adfecdc6c9decba6ca72258536fb2122f
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>